Issue KN-10 fix: Added Test for AuditHistoryActor
